update: move request accessors to their respective feature folder
