+++ src/AuthServer/EndSession/EndSessionRequestAccessor.cs
@@ -3,7 +3,7 @@ using AuthServer.Core.Abstractions;
 using AuthServer.Extensions;
 using Microsoft.AspNetCore.Http;
 
-namespace AuthServer.RequestAccessors.EndSession;
+namespace AuthServer.EndSession;
 
 internal class EndSessionRequestAccessor : IRequestAccessor<EndSessionRequest>
 {
